bind data on HomeFragment from api  with out gbs or map

+++ app/src/main/java/com/m3/weathervue/model/WeatherResponse.kt
@@ -203,7 +203,7 @@ data class DailyItem(
 	val windGust: Any? = null,
 
 	@field:SerializedName("dt")
-	val dt: Int? = null,
+	val dt: Long? = null,
 
 	@field:SerializedName("pop")
 	val pop: Int? = null,
